FossilOrigin-Name: 8ed57d3ad3f8fa0505ccfedc83ed9bcf77d65141b12acc52a88929d822375b53

+++ MSM-Tactical/sys/ush/rensda.sh
@@ -1,6 +1,7 @@
 #!/bin/sh
 set -ex
 cycleda=${1}
+danest=${DANEST:-F}
 head_bv=${HEAD:-bv}
 head_da=${HEAD2:-da}
 ## experiment parameters
@@ -14,6 +15,7 @@ fhour=${INCCYCLE:-6}
 # model resolution
 ires=${IRES:-27}
 # observation settings
+platform=${PLATFORM:-dcdf}
 osse=${OSSE:-F}
 noda=${NODA:-F}
 obsdist=${OBSDIST}
@@ -27,7 +29,7 @@ fixed_level=${DA_FIXED_LEVEL}
 lmin=${LOBSMIN:--60}
 rmin=${ROBSMIN:-60}
 selobs=${SELOBS:-all}
-prep=${PREP:-_preprh}
+prep=${PREP}
 # parallelization
 nisep=${NISEP:-5}
 njsep=${NJSEP:-2}
@@ -78,6 +80,9 @@ export obsdir
 if [ $cycleda -eq 1 ]; then
 fhour=${IOFFSET:-$fhour}
 guesdir=${GUESDIR:-$RUNDIR0}
+if [ $danest = T ]; then
+	fhour=0
+fi
 fi
 bindir=/home/nakashita/Development/grmsm/MSM-Tactical/dpac/build/lmlef
 bindir2=/home/nakashita/Development/grmsm/MSM-Tactical/usr/post
@@ -149,12 +154,22 @@ else
 edate=`date -j -f "%Y%m%d%H%M" -v+${rmin}M +"%H%M" "${adate}00"`
 fi
 if [ $osse = T ];then
+obsin_num=2
 obsf=upper${prep}.siml.${obsdist}.${sdate}-${edate}
 obsf2=surf.siml.${obsdist}.${sdate}-${edate}
 else
+if [ $platform = dcdf ]; then
+obsin_num=2
 obsf=upper${prep}.${sdate}-${edate}
 obsf2=surf.${sdate}-${edate}
 $USHDIR/decode_dcdf.sh $adate $lmin $rmin || exit 9
+elif [ $platform = prepbufr ]; then
+obsin_num=3
+obsf=ADPUPA${prep}.${sdate}-${edate}
+obsf2=ADPSFC.${sdate}-${edate}
+obsf3=SFCSHP.${sdate}-${edate}
+$USHDIR/decode_prepbufr.sh $adate $lmin $rmin || exit 9
+fi
 fi
 ### namelist
 cat <<EOF >lmlef.nml
@@ -169,8 +184,8 @@ cat <<EOF >lmlef.nml
  member=${member},
 &end
 &param_obsope
- obsin_num=2,
- obsin_name='${obsf}','${obsf2}',
+ obsin_num=${obsin_num},
+ obsin_name='${obsf}','${obsf2}','${obsf3}',
  obs_out=F,
  obsout_basename=,
  fguess_basename=,
@@ -232,12 +247,22 @@ EOF
 rm -f ${obsf}.dat ${obsf2}.dat
 ln -s ${obsdir}/${adate}/${obsf}.dat .
 ln -s ${obsdir}/${adate}/${obsf2}.dat .
+if [ $platform = prepbufr ]; then
+rm -f ${obsf3}.dat
+ln -s ${obsdir}/${adate}/${obsf3}.dat .
+fi
 rm -f gues.*.grd ${obsinf}.*.dat ${obsextf}.*.dat
 if [ $mean != T ]; then
 if [ $cycleda -eq 1 ];then
+if [ $danest = T ]; then
+ln -s ${guesdir}/${adate}/${head_da}000/r_sig.f$fh gues.0000.sig.grd
+ln -s ${guesdir}/${adate}/${head_da}000/r_sfc.f$fh gues.0000.sfc.grd
+ln -s ${guesdir}/${adate}/${head_da}000/r_flx.f$fh gues.0000.flx.grd
+else
 ln -s ${guesdir}/${pdate}/r_sig.f$fh gues.0000.sig.grd
 ln -s ${guesdir}/${pdate}/r_sfc.f$fh gues.0000.sfc.grd
 ln -s ${guesdir}/${pdate}/r_flx.f$fh gues.0000.flx.grd
+fi
 else
 ln -s ${guesdir}/${pdate}/${head}000/r_sig.f$fh gues.0000.sig.grd
 ln -s ${guesdir}/${pdate}/${head}000/r_sfc.f$fh gues.0000.sfc.grd
@@ -261,18 +286,30 @@ mem=`printf '%0.3d' $m`
 #if [ $cycleda -eq 1 ] && [ $osse = T ] && [ $m -eq $tmem ]; then
 #else
 mem2=`printf '%0.3d' $em`
+if [ $cycleda -eq 1 ] && [ $danest = T ]; then
+ln -s ${guesdir}/${adate}/${head_da}${mem}/r_sig.f$fh gues.0${mem2}.sig.grd
+ln -s ${guesdir}/${adate}/${head_da}${mem}/r_sfc.f$fh gues.0${mem2}.sfc.grd
+ln -s ${guesdir}/${adate}/${head_da}${mem}/r_flx.f$fh gues.0${mem2}.flx.grd
+else
 ln -s ${guesdir}/${pdate}/${head}${mem}/r_sig.f$fh gues.0${mem2}.sig.grd
 ln -s ${guesdir}/${pdate}/${head}${mem}/r_sfc.f$fh gues.0${mem2}.sfc.grd
 ln -s ${guesdir}/${pdate}/${head}${mem}/r_flx.f$fh gues.0${mem2}.flx.grd
+fi
 if [ $obsda_in = T ];then
 ln -s ${obsdir}/${adate}/${obsextf}.0${mem}.dat ${obsextf}.0${mem2}.dat
 fi
 em=`expr $em + 1`
 if [ $cycleda -eq 1 ] && [ do$psub = doyes ]; then
 mem2=`printf '%0.3d' $em`
+if [ $danest = T ]; then
+ln -s ${guesdir}/${adate}/${head_da}m${mem}/r_sig.f$fh gues.0${mem2}.sig.grd
+ln -s ${guesdir}/${adate}/${head_da}m${mem}/r_sfc.f$fh gues.0${mem2}.sfc.grd
+ln -s ${guesdir}/${adate}/${head_da}m${mem}/r_flx.f$fh gues.0${mem2}.flx.grd
+else
 ln -s ${guesdir}/${pdate}/${head}m${mem}/r_sig.f$fh gues.0${mem2}.sig.grd
 ln -s ${guesdir}/${pdate}/${head}m${mem}/r_sfc.f$fh gues.0${mem2}.sfc.grd
 ln -s ${guesdir}/${pdate}/${head}m${mem}/r_flx.f$fh gues.0${mem2}.flx.grd
+fi
 if [ $obsda_in = T ];then
 ln -s ${obsdir}/${adate}/${obsextf}.m0${mem}.dat ${obsextf}.0${mem2}.dat
 fi
@@ -337,7 +374,7 @@ sed -i -e 's/DATAFILE/'$out'/g' $ctl
 rm ${ctl}-e
 done
 done
-if [ $relax_spread_out = T ];then
+if [ $relax_spread_out = T ] && [ -s rtps.sig.grd ];then
   rm -f fort.*
   ln -s rtps.sig.grd fort.11
   ln -s rtps.bin fort.51
@@ -347,7 +384,7 @@ if [ $relax_spread_out = T ];then
   sed -i -e 's/DATAFILE/'rtps.bin'/g' rtps.ctl
   rm rtps.ctl-e
 fi
-if [ $save_info = T ];then
+if [ $save_info = T ] && [ -s dainfo.sig.grd ];then
 cat << EOF > read_info.nml
 &namlst_info
  member=${member},
@@ -406,9 +443,9 @@ cp r_sig.f00 r_sigitdt
 cp r_sfc.f00 r_sfci
 fi 
 if [ $mean != T ]; then
-if [ -d ${head_da}000 ]; then
-  rm -rf ${head_da}000
-fi
+#if [ -d ${head_da}000 ]; then
+#  rm -rf ${head_da}000
+#fi
 mkdir -p ${head_da}000
 cd ${head_da}000
 ### copy namelists
@@ -433,10 +470,10 @@ fi
 if [ -f ../${obsoutf}.0000.dat ]; then
 mv ../${obsoutf}.0000.dat . 
 fi
-if [ $relax_spread_out = T ];then
+if [ $relax_spread_out = T ] && [ -s ../rtps.sig.grd ];then
 mv ../rtps.* .
 fi
-if [ $save_info = T ];then
+if [ $save_info = T ] && [ -s ../ewgt.ctl ];then
 mv ../dainfo.* .
 mv ../ewgt.ctl .
 fi
@@ -451,9 +488,9 @@ fi
 m=1
 while [ $m -le $member ];do
 mem=`printf '%0.3d' $m`
-if [ -d ${head_da}${mem} ];then
-  rm -rf ${head_da}${mem}
-fi
+#if [ -d ${head_da}${mem} ];then
+#  rm -rf ${head_da}${mem}
+#fi
 mkdir -p ${head_da}${mem}
 cd ${head_da}${mem}
 ### copy namelists
@@ -474,7 +511,7 @@ fi
 if [ -f ../${obsoutf}.0$mem.dat ]; then
 mv ../${obsoutf}.0$mem.dat .
 fi
-if [ $save_info = T ]; then
+if [ $save_info = T ] && [ -f ../ewgt.0$mem.sig.grd ]; then
 mv ../ewgt.0$mem.* .
 fi
 cp r_sig.f00 r_sigi
@@ -489,9 +526,9 @@ for emem in mean sprd;do
 #mv gues.${emem}.sfc.grd ${guesdir}/${pdate}/${head}${emem}/r_sfc.f$fh
 #mv gues.${emem}.bin ${guesdir}/${pdate}/${head}${emem}/
 #mv gues.${emem}.ctl ${guesdir}/${pdate}/${head}${emem}/
-if [ -d ${head_da}${emem} ];then
-  rm -rf ${head_da}${emem}
-fi
+#if [ -d ${head_da}${emem} ];then
+#  rm -rf ${head_da}${emem}
+#fi
 mkdir -p ${head_da}${emem}
 mv gues.${emem}.sig.grd ${head_da}${emem}/r_sig.f$fh
 mv gues.${emem}.sfc.grd ${head_da}${emem}/r_sfc.f$fh
@@ -502,10 +539,10 @@ mv anal.${emem}.sfc.grd ${head_da}${emem}/r_sfc.f00
 mv anal.${emem}.bin ${head_da}${emem}/
 mv anal.${emem}.ctl ${head_da}${emem}/
 if [ $mean = T ] && [ $emem = mean ]; then
-if [ $relax_spread_out = T ];then
+if [ $relax_spread_out = T ] && [ -f rtps.sig.grd ];then
 mv rtps.* ${head_da}${emem}/
 fi
-if [ $save_info = T ];then
+if [ $save_info = T ] && [ -f ewgt.ctl ];then
 mv dainfo.* ${head_da}${emem}/
 mv ewgt.ctl ${head_da}${emem}/
 fi
